test(react): isolate global props event mode tests
