+++ packages/react/runtime/__test__/snapshot/lifecycle/updateGlobalProps.test.jsx
@@ -379,207 +379,4 @@ describe('updateGlobalProps', () => {
       `);
     }
   });
-
-  it('should trigger re-render when useGlobalProps is called', async () => {
-    globalThis.__GLOBAL_PROPS_MODE__ = 'event';
-    const { useGlobalProps, useGlobalPropsChanged, GlobalPropsProvider, GlobalPropsConsumer } = await import(
-      '../../../src/lynx-api?global-props-event'
-    );
-
-    lynx.__globalProps = { theme: 'dark' };
-    let count = 0;
-    let dataTheme, globalPropsTheme;
-    const Comp = () => {
-      const globalProps = useGlobalProps();
-      useGlobalPropsChanged(data => {
-        count++;
-        dataTheme = data.theme;
-        globalPropsTheme = lynx.__globalProps.theme;
-      });
-      return <text>{globalProps.theme}</text>;
-    };
-
-    // main thread render
-    {
-      __root.__jsx = <Comp />;
-      renderPage();
-      expect(count).toBe(0);
-      expect(__root.__element_root).toMatchInlineSnapshot(`
-        <page
-          cssId="default-entry-from-native:0"
-        >
-          <text>
-            <raw-text
-              text="dark"
-            />
-          </text>
-        </page>
-      `);
-    }
-
-    // background render
-    {
-      globalEnvManager.switchToBackground();
-      __root.__jsx = <Comp />;
-      await renderWithCurrentPreact(<Comp />, __root);
-    }
-
-    // hydrate
-    {
-      lynxCoreInject.tt.OnLifecycleEvent(...globalThis.__OnLifecycleEvent.mock.calls[0]);
-    }
-
-    // rLynxChange
-    {
-      globalEnvManager.switchToMainThread();
-      globalThis.__OnLifecycleEvent.mockClear();
-      const rLynxChange = lynx.getNativeApp().callLepusMethod.mock.calls[0];
-      globalThis[rLynxChange[0]](rLynxChange[1]);
-      await waitSchedule();
-    }
-
-    // updateGlobalProps
-    {
-      globalEnvManager.switchToBackground();
-      lynx.getNativeApp().callLepusMethod.mockClear();
-      lynxCoreInject.tt.updateGlobalProps({ theme: 'light' });
-      await waitSchedule();
-
-      // rLynxChange should be called
-      expect(lynx.getNativeApp().callLepusMethod.mock.calls.length).toBe(1);
-      expect(count).toBe(1);
-      expect(dataTheme).toBe('light');
-      expect(globalPropsTheme).toBe('light');
-      // ui change
-      globalEnvManager.switchToMainThread();
-      for (const rLynxChange of lynx.getNativeApp().callLepusMethod.mock.calls) {
-        globalThis[rLynxChange[0]](rLynxChange[1]);
-        rLynxChange[2]();
-      }
-      expect(__root.__element_root).toMatchInlineSnapshot(`
-        <page
-          cssId="default-entry-from-native:0"
-        >
-          <text>
-            <raw-text
-              text="light"
-            />
-          </text>
-        </page>
-      `);
-    }
-  });
-
-  it('should trigger update when GlobalPropsProvider and useGlobalProps are used', async () => {
-    globalThis.__GLOBAL_PROPS_MODE__ = 'event';
-    const { useGlobalProps, useGlobalPropsChanged, GlobalPropsProvider, GlobalPropsConsumer } = await import(
-      '../../../src/lynx-api?global-props-provider-event'
-    );
-
-    lynx.__globalProps = { theme: 'dark' };
-    const Comp = () => {
-      return (
-        <GlobalPropsProvider>
-          <GlobalPropsConsumer>
-            {globalProps => {
-              return <text>{globalProps.theme}</text>;
-            }}
-          </GlobalPropsConsumer>
-        </GlobalPropsProvider>
-      );
-    };
-
-    // main thread render
-    {
-      __root.__jsx = <Comp />;
-      renderPage();
-      expect(__root.__element_root).toMatchInlineSnapshot(`
-        <page
-          cssId="default-entry-from-native:0"
-        >
-          <text>
-            <raw-text
-              text="dark"
-            />
-          </text>
-        </page>
-      `);
-    }
-
-    // background render
-    {
-      globalEnvManager.switchToBackground();
-      __root.__jsx = <Comp />;
-      await renderWithCurrentPreact(<Comp />, __root);
-    }
-
-    // hydrate
-    {
-      lynxCoreInject.tt.OnLifecycleEvent(...globalThis.__OnLifecycleEvent.mock.calls[0]);
-    }
-
-    // rLynxChange
-    {
-      globalEnvManager.switchToMainThread();
-      globalThis.__OnLifecycleEvent.mockClear();
-      const rLynxChange = lynx.getNativeApp().callLepusMethod.mock.calls[0];
-      globalThis[rLynxChange[0]](rLynxChange[1]);
-      await waitSchedule();
-    }
-
-    // updateGlobalProps
-    {
-      globalEnvManager.switchToBackground();
-      lynx.getNativeApp().callLepusMethod.mockClear();
-      lynxCoreInject.tt.updateGlobalProps({ theme: 'light' });
-      await waitSchedule();
-
-      // rLynxChange should be called
-      expect(lynx.getNativeApp().callLepusMethod.mock.calls.length).toBe(2);
-      expect(lynx.getNativeApp().callLepusMethod.mock.calls).toMatchInlineSnapshot(`
-        [
-          [
-            "rLynxChange",
-            {
-              "data": "{"patchList":[{"id":17}]}",
-              "patchOptions": {
-                "flowIds": [
-                  666,
-                ],
-                "reloadVersion": 0,
-              },
-            },
-            [Function],
-          ],
-          [
-            "rLynxChange",
-            {
-              "data": "{"patchList":[{"id":18,"snapshotPatch":[3,-3,0,"light"]}]}",
-              "patchOptions": {
-                "reloadVersion": 0,
-              },
-            },
-            [Function],
-          ],
-        ]
-      `);
-      // ui change
-      globalEnvManager.switchToMainThread();
-      for (const rLynxChange of lynx.getNativeApp().callLepusMethod.mock.calls) {
-        globalThis[rLynxChange[0]](rLynxChange[1]);
-        rLynxChange[2]();
-      }
-      expect(__root.__element_root).toMatchInlineSnapshot(`
-        <page
-          cssId="default-entry-from-native:0"
-        >
-          <text>
-            <raw-text
-              text="light"
-            />
-          </text>
-        </page>
-      `);
-    }
-  });
 });
